Mimic stores previous request client and assigns Request_Client_Mimic
Creating a new mimic instance keeps a record of the previous Request_Client_External::$client and assigns the Request_Client_Mimic to intercept all external request processing. The previous external client is used as the default for Mimic::external_client() but this can be overwritten by configuration file, constructor params or property setters.