Remove match scoring - requests now match first definition
Match scoring is arbitrary and may be hard to predict. Request matching
now simply works through the definition file and matches the first
definition that it finds. Therefore, wildcard definitions must come later
in the definition file than definitions for specific requests.

This is closer in behaviour to Kohana routing, and is considered more
predictable and simpler to maintain. Old behaviour is tagged as
spike-match-scoring in case we want to reintroduce in future.